+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
 Features
 --------
 
+- #200 API circuit endpoints to create, list and search by id or name
+- #195 #196 Enable circuit creation and visualization
 - #172 Add left/right arrow buttons on PC to indicate that the user can change the problem by clicking
 - #168 Add a query builder to ease the handling of MongoDB queries
 - #168 Add API endpoint to get logged in user preferences
@@ -45,6 +47,9 @@ Features
 Fixes
 -----
 
+- #198 Fix hole numbering when undoing a circuit move
+- #189 Fix section and gym not being shown on some ticklist problems
+- Enable CORS on API endpoints
 - #185 Fix downloading a problem as an image 
 - #184 Spacing between radio button and text in hold selection options when creating a problem
 - #183 Display correct text in hold detection text label on the problem creation page